refactor(nexus): add context cancellation and waitgrouping to connection
