+++ nexus/cmd/nexus/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"flag"
-	// "fmt"
 	log "github.com/sirupsen/logrus"
 	"github.com/wandb/wandb/nexus/pkg/server"
 )
@@ -15,12 +14,10 @@ func main() {
 	serveSock := flag.Bool("serve-sock", false, "debug")
 	serveGrpc := flag.Bool("serve-grpc", false, "debug")
 
-	// fmt.Println("Starting 0.0\n")
 	flag.Parse()
 
 	server.InitLogging()
 
-	// log.Println("got", *portFilename, *pid, *debug, *serveSock, *serveGrpc)
 	log.WithFields(log.Fields{
 		"fname":     *portFilename,
 		"pid":       *pid,
